% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structmesh-package.R
\docType{package}
\name{structmesh-package}
\alias{structmesh}
\alias{structmesh-package}
\title{structmesh: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
