% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structuredmesh-package.R
\docType{package}
\name{structuredmesh-package}
\alias{structuredmesh}
\alias{structuredmesh-package}
\title{structuredmesh: Build a Structured Mesh of Quads or Triangles}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
